+++ src/emsuart.h
@@ -1,6 +1,8 @@
 /*
  * emsuart.h
+ * 
  * Header file for emsuart.cpp
+ * 
  * Paul Derbyshire - https://github.com/proddy/EMS-ESP
  */
 #pragma once